refactor!: change Preparation to Class Library, Move ALL enum types from their Interface/GameObj files to a new file EnumType.cs, Add Constant.cs.
