+++ logic/.editorconfig
@@ -1,0 +1,4 @@
+﻿[*.cs]
+
+# CS8632: 只能在 "#nullable" 注释上下文内的代码中使用可为 null 的引用类型的注释。
+dotnet_diagnostic.CS8632.severity = suggestion
